[DX] Extract InvalidItemException
