+++ src/Pim/Bundle/BaseConnectorBundle/Exception/DuplicateProductValueException.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\BaseConnectorBundle\Exception;
 
-use Akeneo\Bundle\BatchBundle\Item\InvalidItemException;
+use Akeneo\Component\Batch\Item\InvalidItemException;
 use Pim\Bundle\TransformBundle\Exception\ParametrizedException;
 use Pim\Bundle\TransformBundle\Exception\ParametrizedExceptionInterface;
