+++ src/Pim/Bundle/TransformBundle/Transformer/AssociationTransformer.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\TransformBundle\Transformer;
 
-use Akeneo\Bundle\BatchBundle\Item\InvalidItemException;
+use Akeneo\Component\Batch\Item\InvalidItemException;
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Pim\Bundle\TransformBundle\Transformer\ColumnInfo\ColumnInfoTransformerInterface;
 use Pim\Bundle\TransformBundle\Transformer\Guesser\GuesserInterface;
